tools/requirements.txt
pyserial>=3.5
matplotlib>=3.7
paho-mqtt>=1.6
